requirements.txt
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
Deprecated==1.2.13
idna==3.4
ifaddr==0.2.0
librespot==0.0.6
mutagen==1.46.0
packaging==21.3
protobuf==3.20.1
pycryptodomex==3.15.0
PyOgg==0.6.14a1
pyparsing==3.0.9
python-dotenv==1.0.0
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.21.0
urllib3==1.26.12
websocket-client==1.4.1
wrapt==1.14.1
zeroconf==0.39.4
